modules/: Move all modules into new dir, rename geolocation.js
